requirements/base.txt
django == 2.1.4
pymysql

djangorestframework

# django跨域

django-cors-headers

djangorestframework-jwt

